+++ util/jtools.js
@@ -214,6 +214,7 @@ E.split_debug_flags = opt=>{
         case '--tls-max-v1.2':
         case '--tls-min-v1.1':
         case '--tls-min-v1.0':
+        case '--insecure-http-parser':
         case '--expose-internals':
         case '--inspect':
         case '--inspect-brk':